pyproject.toml
[tool.poetry]
name = "read-mode"
version = "0.1.0"
description = ""
authors = ["meirdev <maornet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
readability-lxml = "^0.8.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
transformers = {extras = ["torch"], version = "^4.30.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
